backend/pyproject.toml
[tool.poetry]
name = "backend_podcast"
version = "0.1.0"
description = ""
authors = ["ste <stefano.masuelli1999@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.4"
pydantic = "^2.9.2"
uvicorn = "^0.32.0"
peewee = "^3.17.8"
psycopg2-binary = "^2.9.10"
argon2-cffi = "^23.1.0"
pyjwt = "^2.10.0"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
start = "backend_podcast.main:start"
start_docker = "docker.docker:start"
stop_docker = "docker.docker:stop"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
